+++ packages/core/src/polynomial/quadratic.internal.ts
@@ -260,6 +260,9 @@ export const range = dual<
 })
 
 /** @internal */
+export const unitRange = (p: QuadraticPolynomial): Interval.Closed => range(p, Interval.unit)
+
+/** @internal */
 export const length = dual<
   (domain: Interval.Interval) => (p: QuadraticPolynomial) => number,
   (p: QuadraticPolynomial, domain: Interval.Interval) => number